SCB-527 adjust pojo structure
