+++ archetypes/business-service-pojo/src/main/resources/archetype-resources/src/main/java/Hello.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.archetypes;
+package ${groupId};
 
 public interface Hello {
   String hello();
