+++ archetypes/business-service-pojo/src/main/resources/archetype-resources/src/main/resources/microservice.yaml
